tools/scorebot/requirements.txt
slackbot==0.4.1
GitPython==2.1.5
